+++ src/data/defaults.js
@@ -2,7 +2,7 @@
 //  - By Bjorn Stromberg (@bjornstar)
 
 var defaultSettings = {
-	'version': '0.6.1',
+	'version': '0.7.0',
 	'shuffle': false,
 	'repeat': true,
 	'listBlack': [