+++ HISTORY.md
@@ -1,5 +1,10 @@
 #TumTaster Changelog
 
+##v0.7.0
+ * Made links to Tumblr, SoundCloud and Download for each track in the popup.
+ * The track that is playing is highlighted in the popup.
+ * Fixed an issue where posts wouldn't get their download links.
+
 ##v0.6.1
  * Fixed an issue where posts that reappeared would not get their download links reapplied
  * Added this HISTORY file